+++ client/RecoverSecretStruct.rs
@@ -1,0 +1,9 @@
+pub struct RecoverSecretInput {
+    pub word_count: usize,
+    pub letters: String,
+    pub tuple_sizes: Vec<usize>,
+}
+
+pub struct RecoverSecretOutput {
+    pub secret_sentence: String,
+}
